fallacy-detection-service/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
pydantic==2.9.0
python-dotenv==1.0.1
httpx==0.27.0
torch>=2.6.0
transformers>=4.40.0
sentencepiece>=0.2.0
openai>=1.35.0
numpy>=1.26.0
scikit-learn>=1.5.0
datasets>=2.19.0
sentence-transformers>=2.7.0

